fix(api): API요청 인터셉터 Promise 타입으로 변경 및 side effect 수정

+++ Front/src/utils/api/interceptor.ts
@@ -15,54 +15,17 @@ export async function axiosRequestHandler(
         const statusText = error.response?.statusText
         const message = error.response?.data?.messagee
 
-        console.error('Custom Error: ', message)
-
-        if (axios.isAxiosError(error)) {
-            console.error('네트워크 혹은 서버 연결에 문제가 발생했습니다.')
-        }
-
-        if (statusCode === 401 && message === '만료된 access 토큰입니다') {
-            if (userStore.getState().accessToken === '') {
-                /**
-                 * @beta
-                 * @todo 로그아웃 화면 전환
-                 */
-                console.log(111111)
-                return await logout()
-            } else {
-                /**
-                 * @beta
-                 * @todo
-                 * Access Token 파기
-                 * reIssue
-                 * 기존 API 재요청
-                 */
-                console.log(222222)
-                let tempAccessToken = userStore.getState().accessToken
-                userStore.getState().destroyAccessToken()
-                reIssue(tempAccessToken)
-                return await request(...params)
-            }
-        } else if (
+        console.error('Custom Error: ', error)
+        if (
             statusCode === 401 &&
-            message === '만료된 refresh 토큰입니다'
+            message === '만료된 access 토큰입니다' &&
+            userStore.getState().accessToken !== ''
         ) {
-            /**
-             * @beta
-             * @todo 로그아웃
-             */
-            console.log(333333)
-            return await logout()
-        } else if (
-            statusCode === 401 &&
-            message === 'S3 접근 권한이 없습니다.'
-        ) {
-            /**
-             * @beta
-             * @todo S3 접근 권한에 대한 화면 전환
-             */
+            let tempAccessToken = userStore.getState().accessToken
+            userStore.getState().destroyAccessToken()
+            reIssue(tempAccessToken)
+            return await request(...params)
         } else {
-            console.log(444444)
             return await logout()
         }
